little fixes

+++ VetHybPiper.sh
@@ -1,11 +1,17 @@
 ## Bash script to use TargetVet to detect paralogs from data for many samples that have been assembled by HybPiper
+## ARGUMENTS
+#  -V **absolute** directory of VetTargets code: without trailing "/"
+#  -D **absolute** directory of HybPiper output (i.e. with separate folders for each sample and for genes therein): without trailing "/"
+#  -T targets fasta (nucleotide): must be in directory specified in -D
+#  -S file listing sample names to process: must be in directory specified in -D
+#  -G file listing gene names to process: must be in directory specified in -D
+#  -L minimum length of blast matches to keep for analysis
 
 
 ## set defaults
-threads=1
 LENGTH=100
 ## parse args
-while getopts V:D:T:S:G:N:L: option
+while getopts V:D:T:S:G:L: option
 do
 case "${option}"
 in
@@ -15,7 +21,6 @@ D) DIR=${OPTARG};;
 T) TARGETS=${OPTARG};;
 S) SAMPLES=${OPTARG};;
 G) GENES=${OPTARG};;
-N) threads=${OPTARG};;
 L) LENGTH=${OPTARG};;
 
 esac
@@ -24,31 +29,45 @@ done
 # 1. make TargetVet results folder and collate spades scaffolds per sample
 mkdir -p ${DIR}/TargetVet_results/assemblies_collated
 while read i;do
+ FILE=${DIR}/TargetVet_results/assemblies_collated/${i}_all_contigs.fasta
+ if [[ -f "$FILE" ]]; then
+    echo "collated contigs exist for ${i}. Skipping."
+ else
     while read g; do
         if [[ -f ${DIR}/${i}/${g}/${g}_contigs.fasta.gz ]]; then
             gzip -d ${DIR}/${i}/${g}/${g}_contigs.fasta.gz
         fi
         cat ${DIR}/${i}/${g}/${g}_contigs.fasta
-    done < ${GENES} > ${DIR}/TargetVet_results/assemblies_collated/${i}_all_contigs.fasta
-done < ${SAMPLES}
+    done < ${DIR}/${GENES} > ${DIR}/TargetVet_results/assemblies_collated/${i}_all_contigs.fasta
+ fi
+done < ${DIR}/${SAMPLES}
 
 # 2. blast to TARGETS
-mkdir ${DIR}/TargetVet_results/blast_out
+mkdir -p ${DIR}/TargetVet_results/blast_out
 while read i;do
+ FILE=${DIR}/TargetVet_results/blast_out/blastn_`basename ${TARGETS}`_to_${i}_all_contigs.txt
+ if [[ -f "$FILE" ]]; then
+    echo "blast output exists for ${i}. Skipping."
+ else
     blastn -query ${TARGETS} \
         -subject ${DIR}/TargetVet_results/assemblies_collated/${i}_all_contigs.fasta \
-        -out ${DIR}/TargetVet_results/blast_out/blastn_${TARGETS}_to_${i}_all_contigs.txt \
+        -out ${DIR}/TargetVet_results/blast_out/blastn_`basename ${TARGETS}`_to_${i}_all_contigs.txt \
         -evalue 1e-6 \
-        -outfmt "6 qseqid sseqid pident length mismatch gapopen qlen qstart qend slen sstart send evalue bitscore" \
-        -num_threads ${threads}
-done < ${SAMPLES}
+        -outfmt "6 qseqid sseqid pident length mismatch gapopen qlen qstart qend slen sstart send evalue bitscore"
+ fi
+done < ${DIR}/${SAMPLES}
 
 # 3. run VetTargets_genome.R with --doPlots FALSE
-mkdir ${DIR}/TargetVet_results/VetTargets_genome_output
+mkdir -p ${DIR}/TargetVet_results/VetTargets_genome_output
 cd ${DIR}/TargetVet_results/VetTargets_genome_output
 while read i;do
-    BL=${DIR}/TargetVet_results/blast_out/blastn_${TARGETS}_to_${i}_all_contigs.txt
-    BLH=${DIR}/TargetVet_results/blast_out/blastn_${TARGETS}_to_${i}_all_contigs.withHeader.txt
+ FILE=${i}_CoverageStats_AcrossChromosomes.txt
+ if [[ -f "$FILE" ]]; then
+    echo "VetTargets_genome output exists for ${i}. Skipping."
+ else
+    echo "running VetTargets_genome on ${i}"
+    BL=${DIR}/TargetVet_results/blast_out/blastn_`basename ${TARGETS}`_to_${i}_all_contigs.txt
+    BLH=${DIR}/TargetVet_results/blast_out/blastn_`basename ${TARGETS}`_to_${i}_all_contigs.withHeader.txt
     echo -e "qseqid\tsseqid\tpident\tlength\tmismatch\tgapopen\tqlen\tqstart\tqend\tslen\tsstart\tsend\tevalue\tbitscore" | \
         cat - ${BL} > ${BLH}
     
@@ -60,7 +79,8 @@ while read i;do
     --max_intron_percent 1 \
     --min_display_intron 1 \
     --doPlots FALSE > ${i}.Rout
-done < ${SAMPLES} 
+ fi
+done < ${DIR}/${SAMPLES}
 
 # 4. collate ContigStats_acrossChromosomes.txt
 # AND
@@ -68,4 +88,4 @@ done < ${SAMPLES}
 # AND
 # 6. Output genelists for paralogs and single-copy
 
-Rscript ${VETDIR}/DetectParalogs.R -s ${SAMPLES} -d ${DIR}/TargetVet_results/VetTargets_genome_output
+Rscript ${VETDIR}/DetectParalogs.R -s ${DIR}/${SAMPLES} -d ${DIR}/TargetVet_results/VetTargets_genome_output > ${DIR}/TargetVet_results/VetTargets_genome_output/DetectParalogs.Rout
